Remove the need to use the `compat` rules_jvm_external deps
Modern usage of `rules_jvm_external` places everything in a `@maven`
workspace. The compat names were left as a mechanism to ensure
backwards-compatibility, but there's no need to carry on doing this.

This PR removes the use of the compat namespaces and replaces them
with the equivalents from the `@maven` workspace.