Use the `artifact` macro for loading maven deps
The recommended way to load dependencies from `rules_jvm_external`
is to make use of the `@maven` workspace, and the most readable
way of doing that is to use the `artifact` macro provides.

This removes the need to generate the "compat" namespaces, which
`rules_jvm_external` provided for backwards compatibility with
older releases. This change also sets things up for supporting
`bzlmod`: this requires all workspaces accessed by a library to
be named "up front" in the `MODULE.bazel` file. This way, the
only repo that needs to be exported is `@maven`, rather than the
current huge list.

Finally, this PR introduces a lock file for `rules_jvm_external`
which improves local build times by avoiding the need to do a
local resolution. In order to avoid the common failure case of
"add a dep, forget to regenerate the lock file", the
`fail_if_repin_required` attribute is set: builds will fail if
the deps have been updated but the lock file hasn't been.